Use default maplev-config in maplev-view-mode.
If the optional CONFIG argument is missing,
use maplev-config-default to initialize maplev-config.

This is a workaround to allow mds-ss-mode to be derived from
maplev-view-mode.  mds-ss-mode may not have access to maplev-config.

maplev-view-mode uses maplev-config to access :maple.  The better
solution would be to redesign this so that any calls for help or proc
display are passed to the Maple kernel being debugged, however, doing so
is difficult and not ideal.  Serving help pages during a debug session
is questionable, though serving Maple procedures is useful and ensures
that any session-dependent procedures are properly displayed.

A simpler redesign may be for the Maple client to pass the location of
the tty maple executable to the emacs server when it connects; the
server would then use it to create a local maplev-config.  However, if
the server and client are on different machines it may not be practical
to access maple.